Control the hand
I reconstruct the code a little bit. The code is runnable and can
achieve the function we discussed tonight.

p.s. There is a known bug with the current sprite builder. It might
fail if you directly open the ccbproj file using sprite builder.
Instead, open the sprite builder first and then choose to open the
corresponding file.